+++ morpheus-graphql-app/src/Data/Morpheus/App/Internal/Resolving.hs
@@ -41,21 +41,16 @@ module Data.Morpheus.App.Internal.Resolving
     EventHandler (..),
     requireObject,
     ResolverValueDefinition (..),
-    mkObject',
-    mkTypeResolver,
-    mkTypeResolverRef,
-    mkTypeResolvers,
-    getArgument,
-    RelTypeResolver,
-    TypeResolver (..),
+    mkObject'
   )
 where
+import Relude 
 
 import Data.Morpheus.App.Internal.Resolving.Event
 import Data.Morpheus.App.Internal.Resolving.Resolver
 import Data.Morpheus.App.Internal.Resolving.ResolverState
 import Data.Morpheus.App.Internal.Resolving.ResolverValue
 import Data.Morpheus.App.Internal.Resolving.RootResolverValue
-import Data.Morpheus.App.Internal.Resolving.TypeResolvers
 import Data.Morpheus.App.Internal.Resolving.Utils
 import Data.Morpheus.Internal.Ext
+