+++ morpheus-graphql-app/changelog.md
@@ -1,3 +1,7 @@
 # Changelog
 
+## Unreleased Changes
+
+- named resolvers
+
 ## 0.17.0 (Initial Release) - 25.02.2021